requirements.txt
bonobo==0.6.4
psycopg2-binary==2.8.5
pydantic==1.6.1
python-dotenv==0.14.0
requests==2.24.0
selenium==3.141.0
SQLAlchemy==1.3.19
webdriver-manager==3.2.2
